+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## To Be Released
 
+* feat(service): Add optional shard field when registering hosts
+* feat(service): Add `GetForShard` to query hosts for a specific shard
+* feat(service): Add `QueryOptions` to `Service.All/First/One/URL` for shard filtering
+
+Breaking Changes:
+* `Service.All/First/One/URL` signatures now require `QueryOptions`
+* `Service.All/First/One/URL` now return an error and `nil` hosts when no host was found in etcd, instead of returning an empty host
+
 ## v7.1.5
 
 * build(go): use go 1.24